Adds time window to 'New' vods, adds 'Not yet watched'
The new vods list would with time become quite unwieldy. It is now
limited to a 30 day window (configurable) and not yet seen
(configurable) vods.

There is also a new entry level item called 'Not yet watched' which is
your old unseen vods albeit presented through as tree instead of a flat
list.