fixes broken start offset update on match item
For some reason the update works on EntryPage.qml as well even though no
hooks on RecentlyWatchedVideoView.qml are in place to update the start
offset on the view items. I suspect this is due to the fact that the
recently watched videos part on EntryPage.qml flickers oddly on page
activation.