Always make thunbmail available for display
If the app is offline, show thumbnail file, else fallback image.
If the app is online and there is meta data, download the thumbnail if
unavailable. In case there is no meta data, download meta data, then
fetch thumbnail.
If the device is offline and then is connected to the internet, download
thumbnail as if online.

The above should ensure the reasonable behavior a user can expect from
the app. That is download the thumbnail (and whatever data is necessary)
if online, else show fallback.