sc2casts.com scraper: fix pruning of invalid vods
Vods that were added during level 0 parsing but never made it to level 1
page fetch because the fetch would be aborted due to a vod existing in
the database (remember, sc2casts orders strictly by date), had some
valid data in the record and would thus pass the record.valid != 0 check.
I have changed the check to test for a valid url instead.